Added retry support for rate limit exceeded messages. Added support to re-authenticate when the bearer token expires.

+++ src/main/java/com/satalyst/powerbi/model/parsers/NumberColumnTypeParser.java
@@ -19,6 +19,6 @@ public class NumberColumnTypeParser implements ColumnTypeParser {
 
     @Override
     public Object parse(Object o) {
-        return ((Number) o).intValue();
+        return ((Number) o).longValue();
     }
 }
